Move tests to tests/ and install nose
