+++ dev-requirements.txt
@@ -1,6 +1,7 @@
 attrs==17.4.0
 EditorConfig==0.12.1
 mypy==0.560
+nose==1.3.7
 pep8==1.7.1
 pluggy==0.6.0
 psutil==5.4.3
