--- run_tests.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-for i in test_*.py
-do
-    python "$i"
-done
-
